backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
pandas
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
